% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{applicants.us}
\alias{applicants.us}
\title{Given names of SSA applicants.}
\format{A data frame with 4 variables: \code{year}, \code{total}, \code{sex} and
  \code{count}.}
\source{
ssa.gov
}
\usage{
applicants.us
}
\description{
The SSA baby names data comes from social security number (SSN) applications.
}
\keyword{datasets}

