% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{givennames.de}
\alias{givennames.de}
\title{Top Given Names in Germany by year since 1891}
\format{A data frame with 4 variables: \code{name}, \code{rank},
 \code{sex} and \code{year}.}
\source{
beliebte-vornamen.de
}
\usage{
givennames.de
}
\description{
Full baby name data in Germany since 1891, the rank and the gender.
}
\keyword{datasets}

