% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{givennames.fr}
\alias{givennames.fr}
\title{Top Given Names in France by year since 1950.}
\format{A data frame with 4 variables: \code{name}, \code{sex},
 \code{count} and \code{decade}.}
\source{
journaldesfemmes.com
}
\usage{
givennames.fr
}
\description{
Full baby name in France since 1950, the count and the gender.
}
\keyword{datasets}

