% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{givennames.us}
\alias{givennames.us}
\title{Given Names in the US by year since 1880.}
\format{A data frame with 4 variables: \code{name}, \code{sex},
  \code{count} and \code{year}.}
\source{
ssa.gov
}
\usage{
givennames.us
}
\description{
Full baby name data provided by the SSA since 1880, the count and the gender.
}
\keyword{datasets}

